This commit adds permissions to edit the campground by the original author only. Created middleware function to validate author on edit and delete routes of campground
