+++ routes/campground.js
@@ -1,24 +1,9 @@
 const express = require('express')
 const router = express.Router()
 const Campground = require('../models/campground.js');
+const {isLoggedIn, validateCampground, verifyAuthor} = require('../validateMiddleware.js');
+
 const catchAsync = require('../utils/catchAsync')
-const ExpressError = require("../utils/ExpressError");
-const {campgroundSchema} = require("../schemas.js");
-const {isLoggedIn} = require('../validateMiddleware.js');
-const campground = require('../models/campground.js');
-
-
-const validateCampground = (req,res,next) =>{
-    const {error } = campgroundSchema.validate(req.body)
-    if(error){
-        const msg = error.details.map(el => el.message).join(",")
-        throw new ExpressError(msg, 400)
-    } else {
-        next()
-    }
-}
-
-
 
 router.get('/', catchAsync(async (req,res) => {
     const campgrounds = await Campground.find({})
@@ -41,7 +26,7 @@ router.get('/:id', catchAsync(async (req, res) => {
     res.render('campgrounds/show', {campground})
 }))
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async(req,res)=>{
+router.get('/:id/edit', isLoggedIn, verifyAuthor, catchAsync(async(req,res)=>{
     const {id} = req.params
     const findCampgroundById = await Campground.findById(id)
     if(!findCampgroundById){
@@ -62,15 +47,15 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req,res, next
 
 
 
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async(req,res)=>{
+router.put('/:id', isLoggedIn, verifyAuthor, validateCampground, catchAsync(async(req,res)=>{
     const {id} = req.params
     const {campground} = req.body
-    await Campground.findByIdAndUpdate(id, campground, {new: true})
+    const updateCampground = await Campground.findByIdAndUpdate(id, campground, {new: true})
     req.flash('success', 'Successfully updated campground')
-    res.redirect(`/campgrounds/${id}`)
+    res.redirect(`/campgrounds/${updateCampground._id}`)
 }))
 
-router.delete('/:id', isLoggedIn, catchAsync(async (req,res)=> {
+router.delete('/:id', isLoggedIn, verifyAuthor, catchAsync(async (req,res)=> {
     const {id} = req.params
     await Campground.findByIdAndDelete(id)
     req.flash('success', 'Successfully deleted campground')
