+++ routes/review.js
@@ -3,21 +3,10 @@ const router = express.Router({ mergeParams: true})
 
 const Campground = require('../models/campground.js');
 const Review = require('../models/reviews.js')
-
-const { reviewSchema} = require("../schemas.js")
-
+const {validateReview} = require('../validateMiddleware.js')
 const catchAsync = require('../utils/catchAsync')
-const ExpressError = require("../utils/ExpressError");
 
-const validateReview = (req,res, next)=>{
-    const {error} = reviewSchema.validate(req.body)
-    if(error){
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    } else {
-        next()
-    }
-}
+
 
 router.post('/', validateReview, catchAsync(async (req,res) => {
     const {id} = req.params
