+++ validateMiddleware.js
@@ -1,3 +1,7 @@
+const ExpressError = require('./utils/ExpressError');
+const {campgroundSchema, reviewSchema} = require('./schemas');
+const Campground = require('./models/campground')
+
 module.exports.isLoggedIn = (req, res, next) =>{
     if (!req.isAuthenticated()) {
         req.session.returnTo = req.originalUrl
@@ -13,4 +17,34 @@ module.exports.storeReturnTo = (req,res, next) =>{
         res.locals.returnTo = req.session.returnTo
     } 
     next()
+}
+
+module.exports.validateCampground = (req,res,next) =>{
+    const {error } = campgroundSchema.validate(req.body)
+    if(error){
+        const msg = error.details.map(el => el.message).join(",")
+        throw new ExpressError(msg, 400)
+    } else {
+        next()
+    }
+}
+
+module.exports.verifyAuthor = async (req,res,next) => {
+    const {id} = req.params
+    const findCampground = await Campground.findById(id)
+    if (!findCampground.author.equals(req.user._id)){
+        req.flash('error', 'You do not have permission to do that')
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    next()
+}
+
+module.exports.validateReview = (req,res, next)=>{
+    const {error} = reviewSchema.validate(req.body)
+    if(error){
+        const msg = error.details.map(el => el.message).join(',')
+        throw new ExpressError(msg, 400)
+    } else {
+        next()
+    }
 }
